[racl_ctrl,rtl] Use tighter assertion about alert generation
The ASSERT_PRIM_REG_WE_ONEHOT_ERROR_TRIGGER_ALERT assertion isn't
strictly true, which is seen a formal context because the FPV
environment can drive the stateful alert bus so that the alert_p
signal never goes high.

Fortunately, we have separately verified the alert_sender /
alert_receiver pair, so we can change the assertion to just check that
we *ask* the sender to send an alert.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>