[entropy_src/rtl] Fix bucket test result aggregation for Darjeeling
Before this commit, the max bin counter values of the individual
NumBucketHtInst groups (each testing BucketHtDataWidth channels) got
added together. This is not meaningful as it attenuates outliers
observed in individual groups, meaning systematic health test failures
in some channels or groups get buried in noise. This makes it also
harder to come up with useful threshold values as the threshold is not
compared against this final result, but to the results of the individual
NumBucketHtInst groups.

This commit changes the design to instead return the max value of all
NumBucketHtInst groups. This value is most meaningful for defining the
thresholds to be provided by firmware.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>