[csrng/dv] Tweak assertion control in csrng_intr_vseq
This commit addresses a rare dv failure that was caused by turning some
assertions too early on again in csrng_intr_vseq. At the same time, the
commit reduces the window in which assertions get disabled at all to the
part of csrng_intr_vseq which tests the fatal_err interrupt. This is
the only part that uses fault injection, thereby potentially triggering
the assertions in question.

Signed-off-by: Florian Glaser <glaserf@lowrisc.org>