[prim_sha2] Restructured prim_sha2 into separate logical modules
Pulled the compression algorithm logic of SHA2 from prim_sha2 into
a new prim_sha2_compression module to separate the logic from the
SHA2 padding. prim_sha2 now contains the padding and compression
modules with a shift register to load the entire SHA block size
from the padding module to the core. This allows for use of the
SHA core without padding and a FIFO interface, while the default
in HMAC and DMA is FIPS compliant.

Signed-off-by: Evan Apinis <eapinis@zerorisc.com>